let user supply FontName for variable fonts
Decompression becomes a major source of overhead for WOFF fonts, which
means the PostScript name of the font won't be loaded in time.

Eventually we should let just a single OutlineGlyphStore handle all
named instances of a single variable font, just to get the load times
down. Or maybe we should load the system font as a fallback, as long
as we have a way to access these through our facilities.