improve ergonomics for specifying font variation
Named instances are now specified by name instead of by number, and
axes are specified in key-value fashion. The old font variation
specification scheme is still available, as a low-level abstraction,
and for internal use; OutlineFont provides a method to convert from
the more frontend-friendly variation specifier to the old low-level
one, and exclusively use the old low-level specifier for all other
APIs.